Made breakdown budgets clickable
