+++ .flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"firebase_core","path":"/Users/thenamesoak/.pub-cache/hosted/pub.dev/firebase_core-2.7.0/","native_build":true,"dependencies":[]}],"android":[{"name":"firebase_core","path":"/Users/thenamesoak/.pub-cache/hosted/pub.dev/firebase_core-2.7.0/","native_build":true,"dependencies":[]}],"macos":[{"name":"firebase_core","path":"/Users/thenamesoak/.pub-cache/hosted/pub.dev/firebase_core-2.7.0/","native_build":true,"dependencies":[]}],"linux":[],"windows":[],"web":[{"name":"firebase_core_web","path":"/Users/thenamesoak/.pub-cache/hosted/pub.dev/firebase_core_web-2.2.1/","dependencies":[]}]},"dependencyGraph":[{"name":"firebase_core","dependencies":["firebase_core_web"]},{"name":"firebase_core_web","dependencies":[]}],"date_created":"2023-03-06 20:48:55.156127","version":"3.7.0"}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"firebase_core","path":"/Users/thenamesoak/.pub-cache/hosted/pub.dev/firebase_core-2.7.0/","native_build":true,"dependencies":[]}],"android":[{"name":"firebase_core","path":"/Users/thenamesoak/.pub-cache/hosted/pub.dev/firebase_core-2.7.0/","native_build":true,"dependencies":[]}],"macos":[{"name":"firebase_core","path":"/Users/thenamesoak/.pub-cache/hosted/pub.dev/firebase_core-2.7.0/","native_build":true,"dependencies":[]}],"linux":[],"windows":[],"web":[{"name":"firebase_core_web","path":"/Users/thenamesoak/.pub-cache/hosted/pub.dev/firebase_core_web-2.2.1/","dependencies":[]}]},"dependencyGraph":[{"name":"firebase_core","dependencies":["firebase_core_web"]},{"name":"firebase_core_web","dependencies":[]}],"date_created":"2023-03-06 21:01:55.203700","version":"3.7.0"}
