+++ lib/BudgetCategories/expensesPage.dart
@@ -1,0 +1,26 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+
+class ExpensesPage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: "route",
+      home: Scaffold(
+        appBar: AppBar(
+          backgroundColor: Colors.green,
+          title: Text("Test Expense name here"),
+        ),
+        body: Center(
+          child: ElevatedButton(
+              child: Text("Back to Home!"),
+              onPressed: () {
+                Navigator.pop(context);
+              },
+              style: ElevatedButton.styleFrom(
+                  primary: Colors.green)), // RaisedButton
+        ),
+      ),
+    );
+  }
+}
