+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:budgie/button.dart';
 import 'package:flutter/material.dart';
-import 'horProgressBar.dart';
+import 'BudgetCategories/horProgressBar.dart';
 import 'BudgetCircle/overallBudgetCircle.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'firebase_options.dart';
